Add header on indicator list table
