+++ components/indicators/IndicatorList.js
@@ -1,20 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {
-  Badge, Table,
-} from 'reactstrap';
 import { Query } from 'react-apollo';
 import gql from 'graphql-tag';
-import styled from 'styled-components';
 
-import { IndicatorLink } from '../../common/links';
-import Icon from '../common/Icon';
+import { withTranslation } from '../../common/i18n';
 import ContentLoader from '../common/ContentLoader';
 import PlanContext from '../../context/plan';
 import { Meta } from '../layout';
 import ErrorMessage from '../common/ErrorMessage';
 
-import IndicatorListFilters from './IndicatorListFilters';
+import IndicatorListFiltered from './IndicatorListFiltered';
 
 
 const GET_INDICATOR_LIST = gql`
@@ -53,156 +48,6 @@ const GET_INDICATOR_LIST = gql`
   }
 `;
 
-const IndicatorType = styled(Badge)`
-  color: ${(props) => {
-    switch (props.level) {
-      case 'action':
-        return props.theme.themeColors.white;
-      case 'operational':
-        return props.theme.themeColors.black;
-      case 'tactical':
-        return props.theme.themeColors.black;
-      case 'strategic':
-        return props.theme.themeColors.white;
-      default:
-        return props.theme.themeColors.dark;
-    }
-  }};
-  background-color: ${(props) => {
-    switch (props.level) {
-      case 'action':
-        return props.theme.actionColor;
-      case 'operational':
-        return props.theme.operationalIndicatorColor;
-      case 'tactical':
-        return props.theme.tacticalIndicatorColor;
-      case 'strategic':
-        return props.theme.strategicIndicatorColor;
-      default:
-        return '#cccccc';
-    }
-  }};
-`;
-
-const StyledBadge = styled(Badge)`
-  white-space: normal;
-  margin-right: .5em;
-`;
-
-const levels = {
-  operational: { fi: 'toiminnallinen', index: 1 },
-  tactical: { fi: 'taktinen', index: 2 },
-  strategic: { fi: 'strateginen', index: 3 },
-};
-
-class FilteredIndicatorList extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      activeCategory: '',
-      activeSearch: '',
-    };
-    this.handleChange = this.handleChange.bind(this);
-  }
-
-  handleChange(filterType, val) {
-    const change = `active${filterType}`;
-    this.setState({
-      [change]: val,
-    });
-  }
-
-  sortIndicators(indicators) {
-    let sorted = indicators;
-
-    sorted = indicators.sort((a, b) => a.name.localeCompare(b.name));
-    sorted = indicators.sort((a, b) => {
-      if (levels[a.level].index < levels[b.level].index) {
-        return -1;
-      }
-      if (levels[a.level].index > levels[b.level].index) {
-        return 1;
-      }
-      return 0;
-    });
-    return sorted;
-  }
-
-  filterIndicators() {
-    let i;
-    const indicators = this.props.indicators.filter((item) => {
-      const { activeCategory } = this.state;
-      const { activeSearch } = this.state;
-
-      if (activeCategory) {
-        let catMatch = false;
-        for (i = 0; i < item.categories.length; i += 1) {
-          if (item.categories[i].id === activeCategory) catMatch = true;
-        }
-        if (!catMatch) return false;
-      }
-
-      if (activeSearch) {
-        if (item.name.toLowerCase().search(activeSearch.toLowerCase()) !== -1) return true;
-        return false;
-      }
-      return true;
-    });
-
-    return this.sortIndicators(indicators);
-  }
-
-  render() {
-    const indicators = this.filterIndicators();
-
-    return (
-      <div className="mb-5 pb-5">
-        <Meta
-          title={`Mittarit`}
-          description={`Toimenpiteiden edistymistä ja kasvihuonekaasupäästöjen kehitystä seurataan mittareilla`}
-          />
-        <IndicatorListFilters cats={this.props.categories} changeOption={this.handleChange} />
-        <Table hover>
-          <tbody>
-            {indicators.map(item => (
-              <tr key={item.id}>
-                <td>
-                  <IndicatorType pill level={item.level}>
-                    { levels[item.level].fi || <span>-</span> }
-                  </IndicatorType>
-                </td>
-                <td>
-                  <IndicatorLink id={item.id}>
-                    <a>{item.name}</a>
-                  </IndicatorLink>
-                </td>
-                <td>
-                  {item.categories.map(cat => (
-                    <StyledBadge color="light" key={cat.id}>{cat.name}</StyledBadge>
-                  ))}
-                </td>
-                <td>
-                  {(item.latestGraph || item.latestValue) && (
-                    <span>
-                      <Icon name="chartLine" />
-                    </span>
-                  )}
-                </td>
-              </tr>
-            ))}
-          </tbody>
-        </Table>
-      </div>
-    );
-  }
-}
-
-FilteredIndicatorList.propTypes = {
-  indicators: PropTypes.arrayOf(PropTypes.object).isRequired,
-  categories: PropTypes.arrayOf(PropTypes.object).isRequired,
-};
-
-
 class IndicatorList extends React.Component {
   static contextType = PlanContext;
 
@@ -217,7 +62,7 @@ class IndicatorList extends React.Component {
       return indicator;
     });
 
-    let categories = [];
+    const categories = [];
     categoryTypes.forEach((ct) => {
       ct.categories.forEach((cat) => {
         categories.push(cat);
@@ -229,6 +74,7 @@ class IndicatorList extends React.Component {
 
   render() {
     const plan = this.context;
+    const { t } = this.props;
 
     return (
       <Query query={GET_INDICATOR_LIST} variables={{ plan: plan.identifier }}>
@@ -236,11 +82,23 @@ class IndicatorList extends React.Component {
           if (loading) return <ContentLoader />;
           if (error) return <ErrorMessage message={error.message} />;
           const props = this.processDataToProps(data);
-          return <FilteredIndicatorList {...props} />
+          return (
+            <>
+              <Meta
+                title={t('indicators')}
+                description={`Toimenpiteiden edistymistä ja kasvihuonekaasupäästöjen kehitystä seurataan mittareilla`}
+                />
+              <IndicatorListFiltered {...props} />
+            </>
+          )
         }}
       </Query>
     )
   }
 }
 
-export default IndicatorList;
+IndicatorList.propTypes = {
+  t: PropTypes.func.isRequired,
+};
+
+export default withTranslation('common')(IndicatorList);
