+++ static/locales/fi/common.json
@@ -59,7 +59,10 @@
     "indicator-affected-by": "Tähän mittariin vaikuttaa",
     "indicator-has-effect-on": "Tämä mittari vaikuttaa",
     "indicator-not-found": "Mittaria ei löydy",
+    "type": "Tyyppi",
+    "name": "Nimi",
+    "themes": "Teemat",
     "graph": "Kuvaaja",
     "no-graph": "Ei kuvaajaa",
-    "loading": "Ladataan",
+    "loading": "Ladataan"
 }
